+++ clickhouse_connect/driver/httpclient.py
@@ -77,6 +77,7 @@ class HttpClient(Client):
                  https_proxy: Optional[str] = None,
                  server_host_name: Optional[str] = None,
                  apply_server_timezone: Optional[Union[str, bool]] = None,
+                 tz_mode: Optional[str] = None,
                  utc_tz_aware: Optional[Union[bool, Literal["schema"]]] = None,
                  show_clickhouse_errors: Optional[bool] = None,
                  autogenerate_session_id: Optional[bool] = None,
@@ -185,6 +186,7 @@ class HttpClient(Client):
                          query_retries=query_retries,
                          server_host_name=server_host_name,
                          apply_server_timezone=apply_server_timezone,
+                         tz_mode=tz_mode,
                          utc_tz_aware=utc_tz_aware,
                          show_clickhouse_errors=show_clickhouse_errors)
         self.params = dict_copy(self.params, self._validate_settings(ch_settings))