+++ tests/integration_tests/test_timezones.py
@@ -166,7 +166,7 @@ def test_timezone_binding_server(test_client: Client):
     assert server_time.astimezone(pytz.UTC) == utc_time
 
 
-def test_utc_tz_aware(test_client: Client):
+def test_tz_mode(test_client: Client):
     row = test_client.query("SELECT toDateTime('2023-07-05 15:10:40') as dt," +
                             "toDateTime('2023-07-05 15:10:40', 'UTC') as dt_utc",
                             query_tz='UTC').first_row
@@ -175,7 +175,7 @@ def test_utc_tz_aware(test_client: Client):
 
     row = test_client.query("SELECT toDateTime('2023-07-05 15:10:40') as dt," +
                             "toDateTime('2023-07-05 15:10:40', 'UTC') as dt_utc",
-                            query_tz='UTC', utc_tz_aware=True).first_row
+                            query_tz='UTC', tz_mode="aware").first_row
     assert row[0].tzinfo == pytz.UTC
     assert row[1].tzinfo == pytz.UTC
 
@@ -189,7 +189,7 @@ def test_utc_tz_aware(test_client: Client):
 
         row = test_client.query("SELECT toDateTime64('2023-07-05 15:10:40.123456', 6) as dt64," +
                                 "toDateTime64('2023-07-05 15:10:40.123456', 6, 'UTC') as dt64_utc",
-                                query_tz='UTC', utc_tz_aware=True).first_row
+                                query_tz='UTC', tz_mode="aware").first_row
         assert row[0].tzinfo == pytz.UTC
         assert row[1].tzinfo == pytz.UTC
         assert row[0].microsecond == 123456